fix:routers and docs for the Area, Breakdown,Category,Enrollment

+++ app/Http/Controllers/AreaController.php
@@ -17,11 +17,24 @@ use Illuminate\Http\Request;
 class AreaController extends Controller
 {
     /**
-     *
+     * List of areas
+     * @OA\Get(
+     *      path="/areas",
+     *      tags={"Areas"},
+     *      summary="Get list of areas",
+     *      description="Returns list of areas",
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation")
+     *       )
+     *     )
      */
     public function index()
     {
-        return Area::all();
+        return response()->json([
+            'success' =>true,
+            'incidences' => Area::all()
+        ], 200);
     }
 
     /**
@@ -39,25 +52,80 @@ class AreaController extends Controller
     }
 
     /**
-     * Get data of an Area
+     * Get area by id
      *
      * @param int $id
      * @return JsonResponse
      *
+     * @OA\Get (
+     *      path="/area/{id}",
+     *      tags={"Areas"},
+     *      summary="Get a area by id",
+     *      description="Returns the area",
+     *     @OA\Parameter(
+     *          name="id",
+     *          description="Area id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *       ),
+     *      @OA\Response(
+     *          response=404,
+     *          description="The area not be found",
+     *      ),
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      )
+     * )
      */
     public function show(int $id): JsonResponse
     {
         $area=Area::find($id);
 
+        if (!$area) {
+            return response()->json("This area is not exist", '404');
+        }
+
         return response()->json($area, 200) ;
     }
 
     /**
-     * create a new Area
+     * Create a new Area
      * @param Request $request
-     *
      * @return JsonResponse
+     *  * @OA\Post (
+     *      path="/areas",
+     *      tags={"Areas"},
+     *      summary="Create a new area",
+     *      description="Returns created area",
+     *     @OA\Parameter(
+     *          name="request",
+     *          description="request all data",
+     *          required=true,
+     *          in="path",
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
      *
+     *       ),
+     *      @OA\Response(
+     *          response=400,
+     *          description="Bad Request"
+     *      ),
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      )
+     * )
      */
     public function store(Request $request): JsonResponse
     {
@@ -67,17 +135,47 @@ class AreaController extends Controller
     }
 
     /**
-     * Update an  Area
+     * Update the existing area by id
      * @param Request $request
      * @param int $id
      * @return JsonResponse
+     * @OA\Put(
+     *      path="/areas/{id}",
+     *      tags={"Areas"},
+     *      summary="Update a area",
+     *      description="Returns updated area",
+     *     @OA\Parameter(
+     *          name="id",
+     *          description="Area id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
      *
+     *       ),
+     *      @OA\Response(
+     *          response=400,
+     *          description="Bad Request"
+     *      ),
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      )
+     * )
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $parameters = $request->only('name');
 
         $area = Area::find($id);
+        if (!$area) {
+            return response()->json("This area is not exist", '400');
+        }
         $area->name = $parameters['name'];
 
         $area->save();
@@ -86,11 +184,46 @@ class AreaController extends Controller
     }
 
     /**
-     * Delete a Area
+     * Delete the existing area
+     * @param int $id
+     * @return JsonResponse
+     * @OA\Delete  (
+     *      path="/areas/{id}",
+     *      tags={"Areas"},
+     *      summary="Delete a area",
+     *      description="Returns Json response",
+     *     @OA\Parameter(
+     *          name="id",
+     *          description="Area id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
      *
+     *       ),
+     *      @OA\Response(
+     *          response=400,
+     *          description="Bad Request"
+     *      ),
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      ),
+     * )
      */
     public function destroy(int $id): JsonResponse
     {
+
+        $area = Area::find($id);
+
+        if (!$area) {
+            return response()->json("This area is not exist", '400');
+        }
         Area::destroy($id);
 
         return  response()->json('deleted', 200);
